[android] Move conversion code to .cpp files
