+++ platform/android/src/conversion/constant.hpp
@@ -2,14 +2,12 @@
 
 #include "conversion.hpp"
 
-#include <mbgl/util/optional.hpp>
 #include <mbgl/util/color.hpp>
 #include <jni/jni.hpp>
 
 #include <string>
 #include <array>
 #include <vector>
-#include <sstream>
 
 namespace mbgl {
 namespace android {
@@ -17,51 +15,33 @@ namespace conversion {
 
 template <>
 struct Converter<jni::jobject*, bool> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const bool& value) const {
-        static jni::jclass* javaClass = jni::NewGlobalRef(env, &jni::FindClass(env, "java/lang/Boolean")).release();
-        static jni::jmethodID* constructor = &jni::GetMethodID(env, *javaClass, "<init>", "(Z)V");
-        return {&jni::NewObject(env, *javaClass, *constructor, (jboolean) value)};
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const bool& value) const;
 };
 
 template <>
 struct Converter<jni::jboolean, bool> {
-    Result<jni::jboolean> operator()(jni::JNIEnv&, const bool& value) const {
-        return {(jni::jboolean) value};
-    }
+    Result<jni::jboolean> operator()(jni::JNIEnv&, const bool& value) const;
 };
 
 template <>
 struct Converter<jni::jobject*, float> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const float& value) const {
-        static jni::jclass* javaClass = jni::NewGlobalRef(env, &jni::FindClass(env, "java/lang/Float")).release();
-        static jni::jmethodID* constructor = &jni::GetMethodID(env, *javaClass, "<init>", "(F)V");
-        return {&jni::NewObject(env, *javaClass, *constructor, (jfloat) value)};
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const float& value) const;
 };
 
 template <>
 struct Converter<jni::jfloat, float> {
-    Result<jni::jfloat> operator()(jni::JNIEnv&, const float& value) const {
-        return {(jni::jfloat) value};
-    }
+    Result<jni::jfloat> operator()(jni::JNIEnv&, const float& value) const;
 };
 
 
 template <>
 struct Converter<jni::jobject*, double> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const double& value) const {
-        static jni::jclass* javaClass = jni::NewGlobalRef(env, &jni::FindClass(env, "java/lang/Double")).release();
-        static jni::jmethodID* constructor = &jni::GetMethodID(env, *javaClass, "<init>", "(D)V");
-        return {&jni::NewObject(env, *javaClass, *constructor, (jfloat) value)};
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const double& value) const;
 };
 
 template <>
 struct Converter<jni::jdouble, float> {
-    Result<jni::jdouble> operator()(jni::JNIEnv&, const double& value) const {
-        return {(jni::jdouble) value};
-    }
+    Result<jni::jdouble> operator()(jni::JNIEnv&, const double& value) const;
 };
 
 /**
@@ -81,26 +61,17 @@ struct Converter<jni::jobject*, T, typename std::enable_if<std::is_integral<T>::
 
 template <>
 struct Converter<jni::jobject*, std::string> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::string& value) const {
-        return {jni::Make<jni::String>(env, value).Get()};
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::string& value) const;
 };
 
 template <>
 struct Converter<jni::jstring*, std::string> {
-    Result<jni::jstring*> operator()(jni::JNIEnv& env, const std::string& value) const {
-        return {jni::Make<jni::String>(env, value).Get()};
-    }
+    Result<jni::jstring*> operator()(jni::JNIEnv& env, const std::string& value) const;
 };
 
 template <>
 struct Converter<jni::jobject*, Color> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const Color& value) const {
-        std::stringstream sstream;
-        sstream << "rgba(" << value.r << ", " << value.g << ", " << value.b << ", "  << value.a << ")";
-        std::string result = sstream.str();
-        return convert<jni::jobject*, std::string>(env, result);
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const Color& value) const;
 };
 
 template <std::size_t N>
@@ -116,41 +87,19 @@ struct Converter<jni::jobject*, std::array<float, N>> {
 
 template <>
 struct Converter<jni::jobject*, std::vector<std::string>> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::vector<std::string>& value) const {
-        static jni::jclass* stringCass = jni::NewGlobalRef(env, &jni::FindClass(env, "java/lang/String")).release();
-        jni::jarray<jni::jobject>& jarray = jni::NewObjectArray(env, value.size(), *stringCass);
-
-        for(size_t i = 0; i < value.size(); i = i + 1) {
-            Result<jni::jobject*> converted = convert<jni::jobject*, std::string>(env, value.at(i));
-            jni::SetObjectArrayElement(env, jarray, i, *converted);
-        }
-
-        return &jarray;
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::vector<std::string>& value) const;
 };
 
 template <>
 struct Converter<jni::jobject*, std::vector<float>> {
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::vector<float>& value) const {
-        static jni::jclass* floatClass = jni::NewGlobalRef(env, &jni::FindClass(env, "java/lang/Float")).release();
-        jni::jarray<jni::jobject>& jarray = jni::NewObjectArray(env, value.size(), *floatClass);
-
-        for(size_t i = 0; i < value.size(); i = i + 1) {
-            Result<jni::jobject*> converted = convert<jni::jobject*, float>(env, value.at(i));
-            jni::SetObjectArrayElement(env, jarray, i, *converted);
-        }
-
-        return &jarray;
-    }
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const std::vector<float>& value) const;
 };
 
 // Java -> C++
 
 template <>
 struct Converter<std::string, jni::String> {
-    Result<std::string> operator()(jni::JNIEnv& env, const jni::String& value) const {
-        return { jni::Make<std::string>(env, value) };
-    }
+    Result<std::string> operator()(jni::JNIEnv& env, const jni::String& value) const;
 };
 
 } // namespace conversion
