+++ platform/android/src/style/conversion/filter.hpp
@@ -1,30 +1,15 @@
 #pragma once
 
-#include "../android_conversion.hpp"
-#include <mbgl/style/conversion/filter.hpp>
-#include <mbgl/style/conversion_impl.hpp>
+#include <mbgl/style/filter.hpp>
+#include <mbgl/util/optional.hpp>
 
 #include <jni/jni.hpp>
-
-#include <tuple>
-#include <map>
 
 namespace mbgl {
 namespace android {
 namespace conversion {
 
-inline optional<mbgl::style::Filter> toFilter(jni::JNIEnv& env, jni::Array<jni::Object<>> jfilter) {
-    mbgl::optional<mbgl::style::Filter> filter;
-    if (jfilter) {
-      mbgl::style::conversion::Error error;
-      auto converted = mbgl::style::conversion::convert<mbgl::style::Filter>(Value(env, jfilter), error);
-      if (!converted) {
-          mbgl::Log::Error(mbgl::Event::JNI, "Error converting filter: " + error.message);
-      }
-      filter = std::move(*converted);
-    }
-    return filter;
-}
+optional<mbgl::style::Filter> toFilter(jni::JNIEnv&, jni::Array<jni::Object<>>);
 
 } // namespace conversion
 } // namespace android
