+++ platform/android/src/style/conversion/property_expression.hpp
@@ -1,16 +1,11 @@
 #pragma once
 
-#include <mbgl/style/property_value.hpp>
 #include "../../conversion/conversion.hpp"
-#include "../../conversion/constant.hpp"
-#include "types.hpp"
-#include "../../java/lang.hpp"
-
-#include <jni/jni.hpp>
 #include "../../gson/json_element.hpp"
 
-#include <tuple>
-#include <map>
+#include <mbgl/style/property_expression.hpp>
+
+#include <jni/jni.hpp>
 
 namespace mbgl {
 namespace android {
@@ -18,11 +13,8 @@ namespace conversion {
 
 template <class T>
 struct Converter<jni::Object<android::gson::JsonElement>, mbgl::style::PropertyExpression<T>> {
-
     Result<jni::Object<android::gson::JsonElement>> operator()(jni::JNIEnv& env, const mbgl::style::PropertyExpression<T>& value) const {
-        // Convert expressions
-        mbgl::Value expressionValue = value.getExpression().serialize();
-        return gson::JsonElement::New(env, expressionValue);
+        return gson::JsonElement::New(env, value.getExpression().serialize());
     }
 };
