+++ platform/android/src/style/conversion/property_value.hpp
@@ -2,6 +2,7 @@
 
 #include <mbgl/style/color_ramp_property_value.hpp>
 #include <mbgl/style/property_value.hpp>
+
 #include "../../conversion/conversion.hpp"
 #include "../../conversion/constant.hpp"
 #include "property_expression.hpp"
@@ -17,25 +18,22 @@ namespace conversion {
 template <typename T>
 class PropertyValueEvaluator {
 public:
-
     PropertyValueEvaluator(jni::JNIEnv& _env) : env(_env) {}
 
     jni::jobject* operator()(const mbgl::style::Undefined) const {
         return nullptr;
     }
 
-    jni::jobject* operator()(const T &value) const {
-        Result<jni::jobject*> result = convert<jni::jobject*>(env, value);
-        return *result;
+    jni::jobject* operator()(const T& value) const {
+        return *convert<jni::jobject*>(env, value);
     }
 
-    jni::jobject* operator()(const mbgl::style::PropertyExpression<T> &value) const {
-        return *convert<jni::Object<android::gson::JsonElement>, mbgl::style::PropertyExpression<T>>(env, value);
+    jni::jobject* operator()(const mbgl::style::PropertyExpression<T>& value) const {
+        return *convert<jni::Object<android::gson::JsonElement>>(env, value);
     }
 
 private:
     jni::JNIEnv& env;
-
 };
 
 /**
@@ -43,7 +41,6 @@ private:
  */
 template <class T>
 struct Converter<jni::jobject*, mbgl::style::PropertyValue<T>> {
-
     Result<jni::jobject*> operator()(jni::JNIEnv& env, const mbgl::style::PropertyValue<T>& value) const {
         PropertyValueEvaluator<T> evaluator(env);
         return value.evaluate(evaluator);
@@ -55,8 +52,7 @@ struct Converter<jni::jobject*, mbgl::style::PropertyValue<T>> {
  */
 template <>
 struct Converter<jni::jobject*, mbgl::style::ColorRampPropertyValue> {
-
-    Result<jni::jobject*> operator()(jni::JNIEnv& env, const mbgl::style::ColorRampPropertyValue value) const {
+    Result<jni::jobject*> operator()(jni::JNIEnv& env, const mbgl::style::ColorRampPropertyValue& value) const {
         PropertyValueEvaluator<mbgl::style::ColorRampPropertyValue> evaluator(env);
         return *convert<jni::jobject*>(env, value.evaluate(evaluator));
     }
